serverless_deployments/azure_functions/requirements.txt
azure-functions
azure-cosmos
